Updates nginx.dockerfile with permissions changes
Adds arguments and env conversions which pull from docker-compose.yml. Removes the pulled-in nginx.conf file and replaces it with a sed line that changes the default nginx user to the one determined by the env values. Dynamically adds group and user depending on env values, manually exiting as 0 successful (e.g. when using root or www-data users).